Rename folder_Operations to directory_operations
