+++ freqtrade/configuration/directory_operations.py
